Began working on a primitive drawing system and refactored the Camrea system. There is a red dot in the center of the screen now
Former-commit-id: 6cc9009e0f1df1359dfb73980ddf9e857fb1217e [formerly 8497f6e60b5230c86a6f12029e24f88bfe83f2f3 [formerly be0a400ab80261400e7c7cf473ffa4a964bbb41e]]
Former-commit-id: bc35cf807397d1a5aca72589e7061a5ca9dd6068
Former-commit-id: ef958ace82149a22efab4ff7a8380f2c54d1b50e

+++ BrickwareTest/src/Spawner.cpp
@@ -25,9 +25,9 @@ void Spawner::spawnObject()
 	newObject->getTransform()->setPosition(spherePos);
 	newObject->getTransform()->setScale(Vector3(0.1f, 0.1f, 0.1f));
 
-	float randXRot = ((rand() % 314) / 100) + 1;
-	float randYRot = ((rand() % 314) / 100) + 1;
-	float randZRot = ((rand() % 314) / 100) + 1;
+	float randXRot = (float)((rand() % 314) / 100) + 1;
+	float randYRot = (float)((rand() % 314) / 100) + 1;
+	float randZRot = (float)((rand() % 314) / 100) + 1;
 
 	newObject->getTransform()->setRotation(Vector3(randXRot, randYRot, randZRot));
